🎨 Refactor: add path-alias & index.ts
