+++ apps/web/src/app/(feed)/page.tsx
@@ -1,4 +1,4 @@
-import { CommonLayout } from "../../components/common/molecules/CommonLayout";
+import { CommonLayout } from "@/components/common/molecules";
 
 export default function Page(): JSX.Element {
   return <CommonLayout.Contents>web</CommonLayout.Contents>;
