+++ apps/web/src/components/common/icons/Home.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "./type";
+import type { IconProps } from "./@type";
 
 export default function Home({
   width = 24,
