+++ apps/web/src/components/common/icons/Menu.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "./type";
+import type { IconProps } from "./@type";
 
 export default function Menu({
   width = 22,
