+++ apps/web/src/components/common/icons/Mypage.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "./type";
+import type { IconProps } from "./@type";
 
 export default function Mypage({
   width = 24,
