+++ apps/web/src/components/common/icons/Notification.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "./type";
+import type { IconProps } from "./@type";
 
 export default function Notification({
   width = 26,
