+++ apps/web/src/components/common/icons/Search.tsx
@@ -1,4 +1,4 @@
-import type { IconProps } from "./type";
+import type { IconProps } from "./@type";
 
 export default function Search({
   width = 24,
