+++ apps/web/src/components/common/icons/index.ts
@@ -1,0 +1,23 @@
+import Globe from "./Globe";
+import Home from "./Home";
+import Kitty from "./Kitty";
+import Logo from "./Logo";
+import Menu from "./Menu";
+import Mypage from "./Mypage";
+import Notification from "./Notification";
+import Post from "./Post";
+import Search from "./Search";
+import Shorts from "./Shorts";
+
+export {
+  Globe,
+  Home,
+  Kitty,
+  Logo,
+  Menu,
+  Mypage,
+  Notification,
+  Post,
+  Search,
+  Shorts,
+};
