+++ apps/web/src/components/common/molecules/index.ts
@@ -1,0 +1,3 @@
+import { CommonLayout } from "./CommonLayout";
+
+export { CommonLayout };
