+++ apps/web/src/components/common/organisms/index.ts
@@ -1,0 +1,3 @@
+import CommonGbnb from "./CommonGbnb";
+
+export { CommonGbnb };
