+++ apps/web/src/components/feed-tab/atoms/index.ts
@@ -1,0 +1,6 @@
+import ChatButton from "./ChatButton";
+import GlobalButton from "./GlobalButton";
+import LnbMenuButton from "./LnbMenuButton";
+import NotificationButton from "./NotificationButton";
+
+export { ChatButton, GlobalButton, LnbMenuButton, NotificationButton };
