+++ apps/web/src/components/feed-tab/organisms/index.ts
@@ -1,0 +1,3 @@
+import FeedTabHeader from "./FeedTabHeader";
+
+export { FeedTabHeader };
